Modularize the domain search component
Modularize the `DomainSearch` component into separate components for fundamental functionalities.

* **DomainSearch Component**:
  - Create the main `DomainSearch` component in `components/DomainSearch/DomainSearch.tsx`.
  - Manage state and logic for domain search, filtering, sorting, and pagination.
  - Import and use `SearchForm`, `SearchResults`, and `FavoritesList` components.

* **SearchForm Component**:
  - Create the `SearchForm` component in `components/DomainSearch/SearchForm.tsx`.
  - Include input fields, switches, sliders, and buttons for search criteria.
  - Handle form state and trigger search on form submission.

* **SearchResults Component**:
  - Create the `SearchResults` component in `components/DomainSearch/SearchResults.tsx`.
  - Display search results and pagination controls.
  - Handle sorting and pagination logic.

* **FavoritesList Component**:
  - Create the `FavoritesList` component in `components/DomainSearch/FavoritesList.tsx`.
  - Display and manage favorite domains.
  - Handle adding and removing favorites.

* **Remove Old Component**:
  - Delete the `domain-search-component.tsx` file.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/ggindev/domain-pro-check?shareId=XXXX-XXXX-XXXX-XXXX).
